Introduce integer and float rectangles

+++ src/core/entities/entity_sheet.cpp
@@ -2,13 +2,21 @@
 
 namespace albinjohansson::wanderer {
 
-float EntitySheet::GetSourceY(int y, Direction direction) noexcept {
+int EntitySheet::GetSourceY(int y, Direction direction) noexcept {
   switch (direction) {
     default:
-    case Direction::UP: { return static_cast<float>(y); }
-    case Direction::RIGHT: { return static_cast<float>(y + 192); }
-    case Direction::DOWN: { return static_cast<float>(y + 128); }
-    case Direction::LEFT: { return static_cast<float>(y + 64); }
+    case Direction::UP: {
+      return y;
+    }
+    case Direction::RIGHT: {
+      return y + 192;
+    }
+    case Direction::DOWN: {
+      return y + 128;
+    }
+    case Direction::LEFT: {
+      return y + 64;
+    }
   }
 }
